refactor: 未使用の関数、UI、importを削除
- レビュー時のノイズを減らすため、プロジェクト全体から未使用のコード要素を削除
