+++ app/src/main/java/com/tutorial/kneecast/location/GPSLocationManager.kt
@@ -1,16 +1,12 @@
 package com.tutorial.kneecast.location
 
-import android.Manifest
 import android.annotation.SuppressLint
 import android.content.Context
-import android.content.pm.PackageManager
 import android.location.Location
-import androidx.core.app.ActivityCompat
 import com.google.android.gms.location.*
-import com.tutorial.kneecast.MainActivity
 
 @Suppress("DEPRECATION")
-class GPSLocationManager(private val context: Context) {
+class GPSLocationManager(context: Context) {
 
     private var fusedLocationClient: FusedLocationProviderClient =
         LocationServices.getFusedLocationProviderClient(context)
@@ -26,22 +22,6 @@ class GPSLocationManager(private val context: Context) {
     interface MyLocationCallback {
         fun onLocationResult(location: Location?)
         fun onLocationError(error: String)
-    }
-
-    fun getLastLocation(callback: MyLocationCallback) {
-        // 位置情報を取得するためのPermissionチェック
-        if (ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED &&
-            ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-            // 必要な場合、パーミッションを要求
-            if (context is MainActivity) {
-                context.setLocationCallback(callback) // コールバックを保存
-                context.requestLocationPermissions() // 新しいメソッドを使用
-            } else {
-                callback.onLocationError("位置情報の許可が必要です")
-            }
-            return
-        }
-        requestLocationUpdates(callback)
     }
 
     @SuppressLint("MissingPermission")
