+++ app/src/main/java/com/tutorial/kneecast/ui/components/DailyWeatherScreen.kt
@@ -1,7 +1,6 @@
 package com.tutorial.kneecast.ui.components
 
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
