+++ app/src/main/java/com/tutorial/kneecast/ui/components/addressWeather/AddressListItem.kt
@@ -1,6 +1,5 @@
 package com.tutorial.kneecast.ui.components.addressWeather
 
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.icons.Icons
@@ -8,13 +7,11 @@ import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.LocationOn
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import com.tutorial.kneecast.data.model.Feature
-import android.util.Log
 import timber.log.Timber
 
 /**
