+++ app/src/main/java/com/tutorial/kneecast/ui/components/currentLocationWeather/CurrentLocationWeatherView.kt
@@ -1,8 +1,6 @@
 package com.tutorial.kneecast.ui.components.currentLocationWeather
 
-import android.location.Location
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -19,7 +17,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import com.tutorial.kneecast.R
