+++ app/src/main/java/com/tutorial/kneecast/ui/unit/DateFormatter.kt
@@ -1,4 +1,4 @@
-package com.tutorial.kneecast.ui.util
+package com.tutorial.kneecast.ui.unit
 
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
