+++ app/src/main/java/com/tutorial/kneecast/ui/unit/WeatherIconMapper.kt
@@ -1,5 +1,4 @@
-// com/tutorial/kneecast/ui/util/WeatherIconMapper.kt
-package com.tutorial.kneecast.ui.util
+package com.tutorial.kneecast.ui.unit
 
 import com.tutorial.kneecast.R
