+++ app/src/main/java/com/tutorial/kneecast/ui/viewmodel/ViewModelProvider.kt
@@ -3,7 +3,6 @@ package com.tutorial.kneecast.ui.viewmodel
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.tutorial.kneecast.app.AppInitializer
 import com.tutorial.kneecast.ui.LocalAppContext
 import com.tutorial.kneecast.ui.LocalAppInitializer
